Refine placeholders and improve metrics
