+++ examples/distributed/meta_agent_sharding_coordinator_demo.py
@@ -1,0 +1,91 @@
+"""Demonstration of MetaAgentShardingCoordinator with mock components."""
+
+import asyncio
+
+from core.agents.distributed.meta_agent_sharding_coordinator import (
+    MetaAgentShardingCoordinator,
+)
+from packages.core.distributed_inference.model_sharding_engine import (
+    ModelShard,
+    ShardingPlan,
+)
+
+
+async def demo_meta_agent_sharding() -> None:
+    """Run a simple demonstration of the sharding coordinator."""
+    print("\n🔮 Meta-Agent Sharding Coordinator Demo")
+    print("=" * 60)
+
+    # Minimal mock components for demonstration purposes
+    class MockFogCoordinator:
+        pass
+
+    class MockTransportManager:
+        pass
+
+    class MockShardingEngine:
+        async def shard_model(self, model_path, strategy, target_devices):
+            return ShardingPlan(
+                model_name=model_path.split("/")[-1],
+                total_shards=3,
+                shards=[
+                    ModelShard(
+                        shard_id=f"shard_{i}",
+                        device_id=target_devices[i % len(target_devices)],
+                        layer_indices=list(range(i * 8, (i + 1) * 8)),
+                        parameters_count=50_000_000,
+                        memory_mb=200,
+                        compute_requirement=0.8,
+                    )
+                    for i in range(3)
+                ],
+                activation_routing={},
+                memory_efficiency=0.85,
+                compute_balance_score=0.92,
+            )
+
+    class MockDigitalTwin:
+        def get_privacy_report(self):
+            return {"status": "all_local", "data_sources": ["conversations", "location"]}
+
+    coordinator = MetaAgentShardingCoordinator(
+        fog_coordinator=MockFogCoordinator(),
+        transport_manager=MockTransportManager(),
+        sharding_engine=MockShardingEngine(),
+        digital_twin=MockDigitalTwin(),
+    )
+
+    print("\n📋 Creating deployment plan...")
+    target_agents = [
+        "digital_twin_concierge",
+        "king_agent",
+        "magi_agent",
+        "oracle_agent",
+        "sage_agent",
+    ]
+    plan = await coordinator.create_deployment_plan(target_agents)
+    print(f"Plan created with {len(plan.local_agents)} local and {len(plan.fog_agents)} fog agents")
+    print(f"Sharding plans: {list(plan.sharding_plans.keys())}")
+
+    print("\n🚀 Deploying agents...")
+    results = await coordinator.deploy_agents(plan)
+    for agent_name, success in results.items():
+        status = "✅ SUCCESS" if success else "❌ FAILED"
+        print(f"  {agent_name}: {status}")
+
+    print("\n📊 Deployment Status:")
+    status = await coordinator.get_deployment_status()
+    print(f"Local agents: {status['local_agents']}")
+    print(f"Fog agents: {status['fog_agents']}")
+    print(f"Total shards: {status['total_shards']}")
+
+    print("\n🔒 Privacy Report:")
+    privacy = coordinator.get_privacy_report()
+    print(f"Digital twin privacy: {privacy.get('digital_twin_privacy', {}).get('status', 'N/A')}")
+    print(f"Local-only agents: {privacy['data_flows']['local_only']}")
+    print(f"Fog-distributed agents: {privacy['data_flows']['fog_distributed']}")
+
+
+if __name__ == "__main__":
+    asyncio.run(demo_meta_agent_sharding())
+
